# gen-erp-application/public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gen-ERP API",
        "description": "Gen-ERP REST API documentation",
        "contact": {
            "email": "support@gen-erp.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/custom-fields": {
            "get": {
                "tags": [
                    "Custom Fields"
                ],
                "summary": "List all custom fields",
                "operationId": "8b6e1dc464cb8cc3ffcf5856a6f0b779",
                "parameters": [
                    {
                        "name": "entity_type",
                        "in": "query",
                        "description": "Entity type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/CustomField"
                                                    }
                                                ]
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Custom Fields"
                ],
                "summary": "Create a new custom field",
                "operationId": "425b169475b916c513599b5bc87e6165",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "entity_type": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "field_type": {
                                        "type": "string"
                                    },
                                    "required": {
                                        "type": "boolean"
                                    },
                                    "options": {
                                        "type": "array"
                                    },
                                    "default_value": {
                                        "type": "string"
                                    },
                                    "validation_rules": {
                                        "type": "array"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Custom field created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomField"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/custom-fields/{id}": {
            "get": {
                "tags": [
                    "Custom Fields"
                ],
                "summary": "Get a specific custom field",
                "operationId": "93f960b124b11102aa4aad46679b2bfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Custom Field ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomField"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Custom Fields"
                ],
                "summary": "Update a custom field",
                "operationId": "2330c2c565d83a5dc27ff9aa2d80347e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Custom Field ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "field_type": {
                                        "type": "string"
                                    },
                                    "required": {
                                        "type": "boolean"
                                    },
                                    "options": {
                                        "type": "array"
                                    },
                                    "default_value": {
                                        "type": "string"
                                    },
                                    "validation_rules": {
                                        "type": "array"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Custom field updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomField"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Custom Fields"
                ],
                "summary": "Delete a custom field",
                "operationId": "621e57f568d3eb4312c5fcfd24436cc2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Custom Field ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Custom field deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}